requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# Database
sqlalchemy==2.0.23
# sqlite3 is built into Python - removed

# Data Processing
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2

# ML Models
transformers==4.35.2
torch==2.1.1
openai==1.3.8

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2  # Backup for JS-heavy pages

# Text Processing
nltk==3.8.1
textblob==0.17.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
redis==5.0.1  # For caching

# Visualization (for Streamlit)
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0